+++ tests/unit/core/test_etl_dataframe.py
@@ -390,7 +390,7 @@ def test_query(dataframe1, query, expected, as_type):
         raise ValueError(f"Invalid as_type: {as_type}")
 
     # check_index_type=False needed to avoid Attribute "inferred_type" are different
-    assert_frame_equal(result, expected, check_index_type=False)
+    assert_frame_equal(result, expected, check_index_type=len(expected) != 0)
 
 
 def test_query_with_multiple_dicts(dataframe1):